add wasm and ios targets
