+++ gradle.properties
@@ -5,3 +5,5 @@ org.gradle.configuration-cache.parallel=true
 
 org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
 org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true
+
+kotlin.daemon.jvmargs=-Xmx2G
