+++ gradle/libs.versions.toml
@@ -5,8 +5,8 @@ dokka = "2.0.0"
 
 # libraries version
 serialization = "1.7.3"
-coroutines = "1.9.0"
-ktor = "3.0.2"
+coroutines = "1.10.2"
+ktor = "3.1.2"
 mockk = "1.13.13"
 logging = "7.0.0"
 jreleaser = "1.17.0"
@@ -28,7 +28,6 @@ ktor-server-cio = { group = "io.ktor", name = "ktor-server-cio", version.ref = "
 # Testing
 kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
 kotlinx-coroutines-test = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-test", version.ref = "coroutines" }
-kotlinx-coroutines-debug = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-debug", version.ref = "coroutines" }
 ktor-server-test-host = { group = "io.ktor", name = "ktor-server-test-host", version.ref = "ktor" }
 mockk = { group = "io.mockk", name = "mockk", version.ref = "mockk" }
 slf4j-simple = { group = "org.slf4j", name = "slf4j-simple", version.ref = "slf4j" }
