+++ src/commonMain/kotlin/io/modelcontextprotocol/kotlin/sdk/internal/utils.kt
@@ -1,0 +1,5 @@
+package io.modelcontextprotocol.kotlin.sdk.internal
+
+import kotlinx.coroutines.CoroutineDispatcher
+
+internal expect val IODispatcher: CoroutineDispatcher
