+++ src/commonTest/kotlin/io/modelcontextprotocol/kotlin/sdk/shared/ReadBufferTest.kt
@@ -1,21 +1,20 @@
-package shared
+package io.modelcontextprotocol.kotlin.sdk.shared
 
+import io.ktor.utils.io.charsets.Charsets
+import io.ktor.utils.io.core.toByteArray
 import io.modelcontextprotocol.kotlin.sdk.JSONRPCMessage
 import io.modelcontextprotocol.kotlin.sdk.JSONRPCNotification
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
-import io.modelcontextprotocol.kotlin.sdk.shared.ReadBuffer
-import io.modelcontextprotocol.kotlin.sdk.shared.serializeMessage
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertNull
-import org.junit.jupiter.api.Test
-import java.nio.charset.StandardCharsets
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNull
 
 class ReadBufferTest {
     private val testMessage: JSONRPCMessage = JSONRPCNotification(method = "foobar")
 
-    private val json = Json { 
-        ignoreUnknownKeys = true 
+    private val json = Json {
+        ignoreUnknownKeys = true
         encodeDefaults = true
     }
 
@@ -30,13 +29,12 @@ class ReadBufferTest {
         val readBuffer = ReadBuffer()
 
         // Append message without newline
-        val messageBytes = json.encodeToString(testMessage)
-            .toByteArray(StandardCharsets.UTF_8)
+        val messageBytes = json.encodeToString(testMessage).encodeToByteArray()
         readBuffer.append(messageBytes)
         assertNull(readBuffer.readMessage())
 
         // Append newline and verify message is now available
-        readBuffer.append("\n".toByteArray(StandardCharsets.UTF_8))
+        readBuffer.append("\n".encodeToByteArray())
         assertEquals(testMessage, readBuffer.readMessage())
         assertNull(readBuffer.readMessage())
     }
