+++ src/iosMain/kotlin/io/modelcontextprotocol/kotlin/sdk/internal/utils.ios.kt
@@ -1,0 +1,8 @@
+package io.modelcontextprotocol.kotlin.sdk.internal
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.IO
+
+internal actual val IODispatcher: CoroutineDispatcher
+    get() = Dispatchers.IO
