+++ src/jvmMain/java/io/modelcontextprotocol/kotlin/sdk/internal/utils.jvm.kt
@@ -1,0 +1,7 @@
+package io.modelcontextprotocol.kotlin.sdk.internal
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+
+internal actual val IODispatcher: CoroutineDispatcher
+    get() = Dispatchers.IO
