+++ src/wasmJsMain/kotlin/io/modelcontextprotocol/kotlin/sdk/internal/utils.wasmJs.kt
@@ -1,0 +1,7 @@
+package io.modelcontextprotocol.kotlin.sdk.internal
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+
+internal actual val IODispatcher: CoroutineDispatcher
+    get() = Dispatchers.Default
